One more attempt to fix the 'Job failed/done' problem
It seems that sometimes the job-monitoring can produce corrupt output, probably
somewhere in the POE handlers. If this happens occasionally, and not when a job
exits, there is no problem. But if it happens when a file is not in a terminal
state but the job is, that file is then declared as failed!

The solution is to not stop monitoring the job when the job itself is in an
exit-state _unless_ all the files are also in an exit-state. If they're not,
put the job back in for monitoring and go round again.