FTS submission retry value fix
As it is already in the Reporter component the correct value when marking a file as RETRY is 0 not 1. The increasing of the retry count will be done by the RetryManager component. 
In case of value 1, as set if the submission to FTS fails, the retry count will be increased twice, leading to transfer with 3 retries in status RETRY and hence lost by ASO and CRAB.
See: https://hypernews.cern.ch/HyperNews/CMS/get/computing-tools/2951.html